Refactored package names to be more clear an not conflict
