+++ mobile/android/WirelessDebug/app/build.gradle
@@ -26,4 +26,5 @@ dependencies {
     })
     compile 'com.android.support:appcompat-v7:25.2.0'
     testCompile 'junit:junit:4.12'
+    compile project(':wirelessdebugger')
 }
