+++ mobile/android/WirelessDebug/app/src/main/AndroidManifest.xml
@@ -8,7 +8,7 @@
         android:label="@string/app_name"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".app.MainActivity">
+        <activity android:name=".example.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
