+++ mobile/android/WirelessDebug/wirelessdebugger/.gitignore
