+++ mobile/android/WirelessDebug/wirelessdebugger/build.gradle
