+++ mobile/android/WirelessDebug/wirelessdebugger/proguard-rules.pro
