+++ mobile/android/WirelessDebug/wirelessdebugger/src/main/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.google.wireless.debugging">
+    package="com.google.wireless.debugger">
 
     <application android:allowBackup="true" android:label="@string/app_name"
         android:supportsRtl="true">
