+++ mobile/android/WirelessDebug/wirelessdebugger/src/main/java/com/google/wireless/debugger/WirelessDebugger.java
@@ -1,0 +1,7 @@
+package com.google.wireless.debugger;
+
+
+public class WirelessDebugger {
+
+
+}
