+++ mobile/android/WirelessDebug/wirelessdebugger/src/test/java/com/google/wireless/debugger/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.google.wireless.debugging;
+package com.google.wireless.debugger;
 
 import org.junit.Test;
