--- mobile/android/WirelessDebug/wirelessdebugging/src/main/java/com/google/wireless/debugging/WirelessDebugger.java
@@ -1,9 +1,0 @@
-package com.google.wireless.debugging;
-
-
-public class WirelessDebugger {
-
-
-
-
-}
